Finished setting up use of AllDiff constraints and completed issue #6
I hoped the performance were going to improve, but the results are better than what I've anticipated.
The changes cut the time to solve a sudoku up to 80% and I've found no sudoku that takes more than 25 seconds for the most brutal and difficult ones.
For even moderately difficult sudoku it takes less than 5 usually. And all this counting the preliminary inference.